Continue with removal of redundant information
