+++ async-callback-manager/Cargo.toml
@@ -25,3 +25,6 @@ reqwest = { version = "0.12.7", features = [
   "charset",
   "rustls-tls"
 ] }
+
+[lints]
+workspace = true
