--- youtui/src/app/server/messages.rs
@@ -1,1 +1,0 @@
-
